typ: Make Domain a generic type Domain[E]
Add type annotations for foundational types in the domain system. This
makes it at least possible for new code based on the domain system to be
fully typed. A function such as dup_mul_ground could be typed as
```
from sympy.polys.domains.domain import Domain, Er

DUP = list[Er]

def dup_mul_ground(f: DUP[Er], c: Er, K: Domain[Er]) -> DUP[Er]:
    ...
```
This does not attempt to annotate everything in the domain system but
just to ensure that the generic type parameters appear where they need
to be so that it is possible to express the types in generic code.